+++ com.zsmartsystems.zigbee/src/main/java/com/zsmartsystems/zigbee/zcl/clusters/colorcontrol/EnhancedMoveHueCommand.java
@@ -131,8 +131,8 @@ public class EnhancedMoveHueCommand extends ZclColorControlCommand {
 
     @Override
     public void deserialize(final ZclFieldDeserializer deserializer) {
-        moveMode = (Integer) deserializer.deserialize(ZclDataType.ENUMERATION_8_BIT);
-        rate = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+        moveMode = deserializer.deserialize(ZclDataType.ENUMERATION_8_BIT);
+        rate = deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
     }
 
     @Override
