+++ com.zsmartsystems.zigbee/src/main/java/com/zsmartsystems/zigbee/zcl/clusters/colorcontrol/EnhancedMoveToHueCommand.java
@@ -163,9 +163,9 @@ public class EnhancedMoveToHueCommand extends ZclColorControlCommand {
 
     @Override
     public void deserialize(final ZclFieldDeserializer deserializer) {
-        enhancedHue = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
-        direction = (Integer) deserializer.deserialize(ZclDataType.ENUMERATION_8_BIT);
-        transitionTime = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+        enhancedHue = deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+        direction = deserializer.deserialize(ZclDataType.ENUMERATION_8_BIT);
+        transitionTime = deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
     }
 
     @Override
