+++ com.zsmartsystems.zigbee/src/main/java/com/zsmartsystems/zigbee/zcl/clusters/colorcontrol/EnhancedStepHueCommand.java
@@ -156,9 +156,9 @@ public class EnhancedStepHueCommand extends ZclColorControlCommand {
 
     @Override
     public void deserialize(final ZclFieldDeserializer deserializer) {
-        stepMode = (Integer) deserializer.deserialize(ZclDataType.ENUMERATION_8_BIT);
-        stepSize = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
-        transitionTime = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+        stepMode = deserializer.deserialize(ZclDataType.ENUMERATION_8_BIT);
+        stepSize = deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+        transitionTime = deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
     }
 
     @Override
