+++ com.zsmartsystems.zigbee/src/main/java/com/zsmartsystems/zigbee/zcl/clusters/commissioning/RestartDeviceCommand.java
@@ -153,9 +153,9 @@ public class RestartDeviceCommand extends ZclCommissioningCommand {
 
     @Override
     public void deserialize(final ZclFieldDeserializer deserializer) {
-        option = (Integer) deserializer.deserialize(ZclDataType.BITMAP_8_BIT);
-        delay = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_8_BIT_INTEGER);
-        jitter = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_8_BIT_INTEGER);
+        option = deserializer.deserialize(ZclDataType.BITMAP_8_BIT);
+        delay = deserializer.deserialize(ZclDataType.UNSIGNED_8_BIT_INTEGER);
+        jitter = deserializer.deserialize(ZclDataType.UNSIGNED_8_BIT_INTEGER);
     }
 
     @Override
