+++ com.zsmartsystems.zigbee/src/main/java/com/zsmartsystems/zigbee/zcl/clusters/metering/ConfigureNotificationFlags.java
@@ -225,9 +225,9 @@ public class ConfigureNotificationFlags extends ZclMeteringCommand {
 
     @Override
     public void deserialize(final ZclFieldDeserializer deserializer) {
-        issuerEventId = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_32_BIT_INTEGER);
-        notificationScheme = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_8_BIT_INTEGER);
-        notificationFlagAttributeId = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+        issuerEventId = deserializer.deserialize(ZclDataType.UNSIGNED_32_BIT_INTEGER);
+        notificationScheme = deserializer.deserialize(ZclDataType.UNSIGNED_8_BIT_INTEGER);
+        notificationFlagAttributeId = deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
         subPayload = new NotificationCommandSubPayload();
         subPayload.deserialize(deserializer);
     }
